requirements.txt
alembic==1.12.0
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.2
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
gunicorn==22.0.0
itsdangerous==2.1.2
Jinja2==3.1.6
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
pip==23.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==70.0.0
six==1.16.0
SQLAlchemy==2.0.21
SQLAlchemy-serializer==1.4.1
tqdm==4.66.3
typing_extensions==4.8.0
urllib3==2.2.2
Werkzeug==3.0.6
wheel==0.40.0
schedule==1.2.1

